1) Added an Exercise Object and laid out most for an Exercise Object 2) Rethought a Workout Object and laid out a foundation for future coding
